only restart if running

+++ backend/src/manager/mod.rs
@@ -237,10 +237,12 @@ impl Manager {
     pub(super) fn perform_restart(&self) -> impl Future<Output = Result<(), Error>> + 'static {
         let manage_container = self.manage_container.clone();
         async move {
-            let restart_override = manage_container.set_override(MainStatus::Restarting)?;
-            manage_container.wait_for_desired(StartStop::Stop).await;
-            manage_container.wait_for_desired(StartStop::Start).await;
-            restart_override.drop();
+            if manage_container.desired_state().borrow().is_start() {
+                let restart_override = manage_container.set_override(MainStatus::Restarting)?;
+                manage_container.wait_for_desired(StartStop::Stop).await;
+                manage_container.wait_for_desired(StartStop::Start).await;
+                restart_override.drop();
+            }
             Ok(())
         }
     }
